remove factory numba function, only GLRT needs it
